added quads and color to image drawing

+++ src/graphics/graphics.h
@@ -5,8 +5,8 @@
 
 void graphics_init(int width, int height);
 
-void graphics_setBackgroundColor(int red, int green, int blue, int alpha);
-void graphics_setColor(int red, int green, int blue, int alpha);
+void graphics_setBackgroundColor(float red, float green, float blue, float alpha);
+void graphics_setColor(float red, float green, float blue, float alpha);
 void graphics_clear();
 void graphics_swap();
 void graphics_draw_Image(graphics_Image const* image, graphics_Quad const* quad,
